注入对日志接口的实现
