+++ clients/config_client/config_client_test.go
@@ -59,6 +59,7 @@ var serverConfigTest = constant.ServerConfig{
 var serverConfigWithOptions = constant.NewServerConfig("console.nacos.io", 80, constant.WithContextPath("/nacos"))
 
 var clientConfigWithOptions = constant.NewClientConfig(
+	constant.WithLogEntity(mockLogger{}),
 	constant.WithTimeoutMs(10*1000),
 	constant.WithBeatInterval(2*1000),
 	constant.WithNotLoadCacheAtStart(true),
@@ -685,4 +686,39 @@ func TestGetConfigWithSpecialSymbol(t *testing.T) {
 
 	assert.Nil(t, err)
 	assert.Equal(t, contentStr, content)
+}
+
+type mockLogger struct {
+}
+
+func (m mockLogger) Info(args ...interface{}) {
+	panic("implement me")
+}
+
+func (m mockLogger) Warn(args ...interface{}) {
+	panic("implement me")
+}
+
+func (m mockLogger) Error(args ...interface{}) {
+	panic("implement me")
+}
+
+func (m mockLogger) Debug(args ...interface{}) {
+	panic("implement me")
+}
+
+func (m mockLogger) Infof(fmt string, args ...interface{}) {
+	panic("implement me")
+}
+
+func (m mockLogger) Warnf(fmt string, args ...interface{}) {
+	panic("implement me")
+}
+
+func (m mockLogger) Errorf(fmt string, args ...interface{}) {
+	panic("implement me")
+}
+
+func (m mockLogger) Debugf(fmt string, args ...interface{}) {
+	panic("implement me")
 }
