+++ clients/nacos_client/nacos_client.go
@@ -25,7 +25,6 @@ import (
 	"github.com/nacos-group/nacos-sdk-go/common/constant"
 	"github.com/nacos-group/nacos-sdk-go/common/file"
 	"github.com/nacos-group/nacos-sdk-go/common/http_agent"
-	"github.com/nacos-group/nacos-sdk-go/common/logger"
 )
 
 type NacosClient struct {
@@ -76,7 +75,7 @@ func (client *NacosClient) SetClientConfig(config constant.ClientConfig) (err er
 		}
 	}
 
-	logger.Infof("logDir:<%s>   cacheDir:<%s>", config.LogDir, config.CacheDir)
+	//logger.Infof("logDir:<%s>   cacheDir:<%s>", config.LogDir, config.CacheDir)
 	client.clientConfig = config
 	client.clientConfigValid = true
